refactor: path 경로 변경
- 표면 계층에 path 경로 명시
