+++ query-server/src/main/java/project/goorm/queryserver/business/web/news/presentation/NewsQueryByBookmarkWithMemberIdAPI.java
@@ -4,7 +4,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
-import project.goorm.queryserver.business.web.news.application.NewsQueryByBookmarkService;
+import project.goorm.queryserver.business.web.news.application.service.NewsQueryByBookmarkService;
 import project.goorm.queryserver.business.web.news.presentation.response.NewsResponse;
 import project.goorm.queryserver.common.annotation.helper.PresentationLayer;
 import project.goorm.queryserver.common.response.ApiResponse;
@@ -12,7 +12,7 @@ import project.goorm.queryserver.common.response.ApiResponse;
 import java.util.List;
 
 @PresentationLayer
-@RequestMapping("/api/news/bookmark/id")
+@RequestMapping(path = "/api/news/bookmark/id")
 public class NewsQueryByBookmarkWithMemberIdAPI {
 
     private final NewsQueryByBookmarkService newsQueryByBookmarkService;
@@ -21,9 +21,9 @@ public class NewsQueryByBookmarkWithMemberIdAPI {
         this.newsQueryByBookmarkService = newsQueryByBookmarkService;
     }
 
-    @GetMapping("/{member_id}")
-    public ResponseEntity<ApiResponse> findNewsByBookmarkWithMemberId(@PathVariable("member_id") Long memberId) {
-        List<NewsResponse> data = newsQueryByBookmarkService.findNewsByBookmarkWithMemberId(memberId);
+    @GetMapping(path = "/{memberId}")
+    public ResponseEntity<ApiResponse> findNewsByBookmarkWithMemberId(@PathVariable("memberId") Long memberId) {
+        List<NewsResponse> data = newsQueryByBookmarkService.findBookMarkByMemberId(memberId);
         return ResponseEntity.ok()
                 .body(ApiResponse.of(data));
     }
