+++ query-server/src/main/java/project/goorm/queryserver/business/web/news/presentation/NewsQueryByCompanyIdAPI.java
@@ -4,7 +4,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
-import project.goorm.queryserver.business.web.news.application.NewsQueryByCompanyService;
+import project.goorm.queryserver.business.web.news.application.service.NewsQueryByCompanyService;
 import project.goorm.queryserver.business.web.news.presentation.response.NewsResponse;
 import project.goorm.queryserver.common.annotation.helper.PresentationLayer;
 import project.goorm.queryserver.common.response.ApiResponse;
@@ -12,7 +12,7 @@ import project.goorm.queryserver.common.response.ApiResponse;
 import java.util.List;
 
 @PresentationLayer
-@RequestMapping("/api/news/company/id")
+@RequestMapping(path = "/api/news/company/id")
 public class NewsQueryByCompanyIdAPI {
 
     private final NewsQueryByCompanyService newsQueryByCompanyService;
@@ -21,7 +21,7 @@ public class NewsQueryByCompanyIdAPI {
         this.newsQueryByCompanyService = newsQueryByCompanyService;
     }
 
-    @GetMapping("/{companyid}")
+    @GetMapping(path = "/{companyid}")
     public ResponseEntity<ApiResponse> findNewsByCompanyId(@PathVariable("companyid") Long companyId) {
         List<NewsResponse> data = newsQueryByCompanyService.findNewsByCompanyId(companyId);
         return ResponseEntity.ok()
