+++ query-server/src/main/java/project/goorm/queryserver/business/web/news/presentation/NewsQueryByCompanyNameAPI.java
@@ -4,7 +4,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
-import project.goorm.queryserver.business.web.news.application.NewsQueryByCompanyService;
+import project.goorm.queryserver.business.web.news.application.service.NewsQueryByCompanyService;
 import project.goorm.queryserver.business.web.news.presentation.response.NewsResponse;
 import project.goorm.queryserver.common.annotation.helper.PresentationLayer;
 import project.goorm.queryserver.common.response.ApiResponse;
@@ -12,7 +12,7 @@ import project.goorm.queryserver.common.response.ApiResponse;
 import java.util.List;
 
 @PresentationLayer
-@RequestMapping("/api/news/company/name")
+@RequestMapping(path = "/api/news/company/name")
 public class NewsQueryByCompanyNameAPI {
 
     private final NewsQueryByCompanyService newsQueryByCompanyService;
@@ -21,7 +21,7 @@ public class NewsQueryByCompanyNameAPI {
         this.newsQueryByCompanyService = newsQueryByCompanyService;
     }
 
-    @GetMapping("/{companyName}")
+    @GetMapping(path = "/{companyName}")
     public ResponseEntity<ApiResponse> findNewsByCompanyName(@PathVariable("companyName") String companyName) {
         List<NewsResponse> data = newsQueryByCompanyService.findNewsByCompanyName(companyName);
         return ResponseEntity.ok()
